Split stores in different files
