+++ src/pycrdt/store/__init__.py
@@ -1,6 +1,7 @@
-from .store import BaseYStore as BaseYStore
-from .store import SQLiteYStore as SQLiteYStore
-from .store import TempFileYStore as TempFileYStore
-from .store import YDocNotFound as YDocNotFound
+from .base import BaseYStore as BaseYStore
+from .base import YDocNotFound as YDocNotFound
+from .sqlite import SQLiteYStore as SQLiteYStore
+from .file import FileYStore as FileYStore
+from .file import TempFileYStore as TempFileYStore
 
 __version__ = "0.1.0"
